refactor(asv3): split chatStore into threads/streaming/proposals/messages (WP-3)
Decomposes the 648-LOC chatStore into four Pinia stores aligned with its real
lifecycles: chatThreadsStore (persisted), streamingTurnStore (per-turn),
proposalStore (cross-cutting), messagesStore (per-thread message log + chat
panel I/O surface).

Adds useChatReset composable as the single source of truth for "new
conversation"; closes UX-#15 (streamingText not reset on New conversation).

Adds the streaming-reset-on-transport-change test (Testing-F8) plus a
cross-store reset assertion in tests/ui/composables/useChatReset.test.ts.

chatStore.ts deleted per CLAUDE.md no-back-compat policy.

Closes:
- Arch review #4 (carve chatStore into Threads/StreamingTurn/Proposals)
- UX review #15 (streamingText not reset on New conversation)
- Testing review F8 (transport-change reset invariant untested)
