+++ src/ui/stores/proposalStore.ts
@@ -1,0 +1,100 @@
+import { defineStore } from 'pinia';
+import { ref } from 'vue';
+
+import type {
+	FileWriteProposal,
+	FileWriteProposalStatus,
+} from '@/application/chat/FileWriteProposal';
+import type { CommitProposalErrorCode } from '@/application/chat/errors';
+
+/**
+ * Pinia store for **cross-cutting** file-write proposals (Arch review #4,
+ * WP-3 split of the former monolithic `chatStore`).
+ *
+ * Proposals straddle both the persisted-threads and the per-turn lifecycles —
+ * they're scoped to a thread (`proposal.threadId`) but generated by a single
+ * structured turn and may resolve across many turns. Living in their own
+ * store keeps the threads store focused on persistence-shaped data and lets
+ * the message UI render Accept/Reject controls without subscribing to
+ * streaming state.
+ *
+ * REQ-ASM-041, REQ-ASM-043, REQ-ASM-045.
+ */
+export const useProposalStore = defineStore('proposals', () => {
+	/**
+	 * Unresolved or recently-decided file-write proposals, keyed by
+	 * `proposalId`. REQ-ASM-041.
+	 */
+	const proposals = ref<Map<string, FileWriteProposal>>(new Map());
+
+	/**
+	 * Stores a new `FileWriteProposal`. Replaces any prior proposal with the
+	 * same `proposalId` (idempotent re-emission is tolerated). REQ-ASM-041.
+	 */
+	function addProposal(proposal: FileWriteProposal): void {
+		const next = new Map(proposals.value);
+		next.set(proposal.proposalId, proposal);
+		proposals.value = next;
+	}
+
+	/**
+	 * Transitions a proposal to a new lifecycle status. Stamps `decidedAt`
+	 * with "now" for any non-`pending` status. Records `failureReason` only
+	 * when the status transitions to `'failed'`; clears it otherwise. No-op
+	 * if the proposal is unknown. REQ-ASM-043, REQ-ASM-045.
+	 */
+	function setProposalStatus(
+		proposalId: string,
+		nextStatus: FileWriteProposalStatus,
+		failureReason?: CommitProposalErrorCode,
+	): void {
+		const existing = proposals.value.get(proposalId);
+		if (existing === undefined) return;
+		const decidedAt = nextStatus === 'pending' ? null : new Date().toISOString();
+		const next = new Map(proposals.value);
+		next.set(proposalId, {
+			...existing,
+			status: nextStatus,
+			decidedAt,
+			failureReason: nextStatus === 'failed' ? (failureReason ?? null) : null,
+		});
+		proposals.value = next;
+	}
+
+	/**
+	 * Drops every `FileWriteProposal` whose `threadId` matches. Pairs with
+	 * `messagesStore.clearThreadMessages` for the "New conversation" reset
+	 * path — Codex P2 on PR #369: without this, proposals (including their
+	 * `envelope.content` payloads) accumulated unreachably across
+	 * `/create` + reset cycles since there's no thread switcher in
+	 * Increment 1 to revisit them.
+	 */
+	function clearThreadProposals(threadId: string): void {
+		let mutated = false;
+		const next = new Map(proposals.value);
+		for (const [proposalId, proposal] of proposals.value) {
+			if (proposal.threadId === threadId) {
+				next.delete(proposalId);
+				mutated = true;
+			}
+		}
+		if (mutated) proposals.value = next;
+	}
+
+	/**
+	 * Wipe every proposal. Used by `useChatReset.resetAll()` and test
+	 * fixtures. Pinia's default `$reset` does not restore Maps, hence the
+	 * explicit action.
+	 */
+	function reset(): void {
+		proposals.value = new Map();
+	}
+
+	return {
+		proposals,
+		addProposal,
+		setProposalStatus,
+		clearThreadProposals,
+		reset,
+	};
+});
