+++ tests/ui/components/chat/ChatSidebar.sessionPersistence.test.ts
@@ -44,7 +44,7 @@ import {
 	TRANSPORT_KIND_KEY,
 } from '@/infrastructure/bridge/ports'
 import type { TransportKind } from '@/domain/chat/TransportKind'
-import { useChatStore } from '@/ui/stores/chatStore'
+import { getChatStoresFacade } from '../../../__fakes__/chatStoresFacade'
 import { ChatSidebarPO } from './ChatSidebar.po'
 import type { PluginSettings } from '@/domain/settings/PluginSettings'
 import { DEFAULT_SETTINGS } from '@/domain/settings/PluginSettings'
@@ -160,11 +160,11 @@ async function mountSidebar(args: {
 	})
 
 	await flushPromises()
-	const store = useChatStore(pinia)
+	const store = getChatStoresFacade(pinia)
 	return { wrapper, port, bridge, po: new ChatSidebarPO(wrapper), store, transportKindRef }
 }
 
-async function send(store: ReturnType<typeof useChatStore>, po: ChatSidebarPO, text: string) {
+async function send(store: ReturnType<typeof getChatStoresFacade>, po: ChatSidebarPO, text: string) {
 	store.setUserText(text)
 	await flushPromises()
 	await po.clickSend()
