+++ tests/ui/components/chat/ChatSidebar.test.ts
@@ -23,7 +23,7 @@ import {
 	LOGGER_PORT,
 	SECRET_STORE_PORT,
 } from '@/infrastructure/bridge/ports'
-import { useChatStore } from '@/ui/stores/chatStore'
+import { useMessagesStore } from '@/ui/stores/messagesStore'
 import { ChatSidebarPO } from './ChatSidebar.po'
 import type { PluginSettings } from '@/domain/settings/PluginSettings'
 import { DEFAULT_SETTINGS } from '@/domain/settings/PluginSettings'
@@ -107,7 +107,7 @@ async function mountSidebar(options: {
 
 	await flushPromises()
 
-	const store = useChatStore(pinia)
+	const store = useMessagesStore(pinia)
 	return { wrapper, port, bridge, po: new ChatSidebarPO(wrapper), store }
 }
