+++ tests/ui/components/chat/ChatSidebar.threadRotation.test.ts
@@ -31,7 +31,7 @@ import {
 	TRANSPORT_KIND_KEY,
 } from '@/infrastructure/bridge/ports';
 import type { TransportKind } from '@/domain/chat/TransportKind';
-import { useChatStore } from '@/ui/stores/chatStore';
+import { getChatStoresFacade } from '../../../__fakes__/chatStoresFacade';
 import { ChatSidebarPO } from './ChatSidebar.po';
 import type { PluginSettings } from '@/domain/settings/PluginSettings';
 import { DEFAULT_SETTINGS } from '@/domain/settings/PluginSettings';
@@ -92,12 +92,12 @@ async function mountSidebar(resolvedKind: TransportKind = 'api-key') {
 		},
 	});
 	await flushPromises();
-	const store = useChatStore(pinia);
+	const store = getChatStoresFacade(pinia);
 	return { wrapper, store, po: new ChatSidebarPO(wrapper), transportKindRef };
 }
 
 function seedThread(
-	store: ReturnType<typeof useChatStore>,
+	store: ReturnType<typeof getChatStoresFacade>,
 	record: ChatThreadRecord,
 	messageTexts: string[],
 ): void {
